Add food expiring process

+++ Assets/Prototype/Scripts/Items/Item.cs
@@ -10,12 +10,12 @@ namespace Prototype.Scripts.Items
         public bool IsFood => NutritionalValue > 0f;
 
         public string Name;
-
+        
         [Header("For fuel")]
         public int ForgeClickCount;
 
         [Header("For food")] 
         public float NutritionalValue;
-        public float ExpirationTimeSeconds;
+        public int ExpirationTimeSeconds;
     }
 }
